remove webview message handling from plots
